add new hyfeatures test, remove old tests
