+++ pytest.ini
@@ -3,8 +3,7 @@ addopts =
     --ignore=src/troute-network/troute/test_qlat.py 
     --ignore=src/troute-network/troute/network/reservoirs/test/test_compute_kernel.py 
     --ignore=src/troute-routing/troute/routing/test_compute.py
-    --ignore=src/troute-network/tests/test_HYFeaturesNetwork.py
     --ignore=src/troute-config/test/test_config.py
-    --ignore=src/troute-network/tests/test_HYFeaturesNetwork.py
     --ignore=src/kernel/muskingum/test_MC_kernel.py
     --ignore=src/troute-routing/test_framework_parity.py
+    --ignore=src/troute-network/tests/test_HYFeaturesNetwork_needs_refactor.py
