+++ src/troute-network/tests/test_HYFeaturesNetwork.py
@@ -1,118 +1,16 @@
-import pytest
-from pathlib import Path
 import pandas as pd
-from troute.HYFeaturesNetwork import HYFeaturesNetwork
-from troute.routing.compute import compute_nhd_routing_v02
-#set the workdir relative to this test config
-#and use that to look for test data
-_workdir=Path(__file__).parent
+import io
+from troute.HYFeaturesNetwork import HYFeaturesNetwork as hyfn
 
-"""
-Fixtures for setting up various components for testing
-"""
+def test_get_da_gages():
+    usgs_ind_c = 'id,value\n2624217,True\n2624223,False\n2624228,True\n2624235,True\n2624239,True\n2624254,True\n2624266,False\n2624267,True\n2624286,True\n2624289,True\n2624302,False\n2624305,True\n2624323,True\n2624350,True\n2624372,False\n2624374,False\n2624385,False\n2624389,False\n2624409,False\n2624410,True\n2624421,False\n2624423,True\n2624441,True\n2624450,True\n2624462,True\n2624473,True\n2624476,True\n2624499,True\n2624508,True\n2624543,True\n2624563,True\n2624575,False\n2624577,True\n2624610,False\n2624611,True\n2624632,True\n2624650,True\n2624672,False\n2624675,True\n2624682,False\n2624684,True\n2624696,True\n2624741,True\n2624746,True\n2624759,True\n2624779,True\n2624798,True\n2624798,False\n2624824,True\n2624831,False\n2624832,False\n2624832,False\n2624833,False\n2624834,True\n2624835,False\n2624841,False\n2624848,True\n2624849,False\n2624871,False\n2624874,True\n2624894,True\n2624907,False\n2624909,True\n2624941,True\n2624960,True\n2624962,False\n2624963,True\n2624966,True\n2624968,False\n2624970,False\n2624978,True\n2624984,True\n2625000,True\n2625003,True\n2625007,False\n2625008,True\n2625010,False\n2625031,True\n2625037,True\n2625061,False\n2625095,True\n2625107,True\n2625111,True\n2625117,True\n2625227,False\n2625240,True\n2625253,True\n2625253,False\n2625302,True\n2625304,False\n2625316,True\n2625370,True\n2625372,True\n2625373,True\n2625374,True\n2625486,False\n2625487,True\n2625490,True\n2625513,True\n2625514,True\n2625570,True\n2625571,False\n2625649,True\n2625684,True\n2625684,False\n2625685,True\n2625687,True\n2625726,False\n2625749,False\n2625845,True\n2625897,False\n2625902,True\n2625965,False\n2625977,False\n2625994,False\n2626239,True\n2626451,True\n2626481,True\n2626678,True\n2626743,True\n2626747,True\n2626861,True\n2626861,False\n2626900,True\n2626928,True\n2626929,True\n2626947,True\n2627067,False\n2627087,False\n2627163,True\n2627186,True\n2627280,True\n2627295,True\n2627312,True\n2627353,False\n2627354,True\n2627369,True\n2627598,False\n2627628,False\n2627697,False\n2627762,True\n2627786,True\n2627858,True\n2627860,True\n2627862,False\n2627876,True\n2627877,True\n2627877,False\n2627978,True\n2628056,False\n2628058,False\n2628146,True\n2628205,False\n2628214,False\n2628218,False\n2628262,False\n2628262,False\n2628263,False\n2628442,True\n2628447,False\n2628477,True\n2628479,True\n2628577,True\n2628577,False\n2628583,True\n2628584,True\n2628586,True\n2628630,True\n2628633,False\n2628634,False\n2628647,False\n2628819,True\n2628941,True\n2628998,False\n2629009,True\n2629108,True\n2629155,True\n2629188,False\n2629190,False\n2629326,True\n2629540,True\n2629541,True\n2629543,True\n2629544,True\n2629545,True\n2629551,True\n2629596,True\n2629597,True\n2629599,True\n2629620,True\n2629625,True\n2629626,False\n2629627,True\n2629628,True\n2629636,True\n2629668,True\n2629693,True\n2629847,True\n2630271,True\n2630435,True\n2630459,True\n2630464,True\n2630538,True\n2630612,False\n2630690,True\n2630727,True\n2630897,True\n2631093,True\n'
 
-def supernetwork_parameters(type):
-    if( type == 'gpkg'):
-       geo_file =  _workdir.joinpath("data/hy_network.gpkg")
-       edge_list = None
-    elif( type == 'json'):
-        geo_file =  _workdir.joinpath("data/flowpath_attributes.json")
-        edge_list = _workdir.joinpath("data/flowpath_edge_list.json")
-    params = {
-        "title_string":"HY_Features Test",
-        "geo_file_path":geo_file,
-        "flowpath_edge_list":edge_list,
-        "columns": {
-            #link????
-            "key": "id",
-            "downstream": "toid",
-            "dx": "lengthkm",
-            "n": "n",  # TODO: rename to `manningn`
-            "ncc": "nCC",  # TODO: rename to `mannningncc`
-            "s0": "slope_percent",  # TODO: rename to `bedslope`
-            "bw": "BtmWdth",  # TODO: rename to `bottomwidth`
-            #waterbody: "NHDWaterbodyComID",
-            "tw": "TopWdth",  # TODO: rename to `topwidth`
-            "twcc": "TopWdthCC",  # TODO: rename to `topwidthcc`
-            #alt: "alt",
-            "musk": "MusK",
-            "musx": "MusX",
-            "cs": "ChSlp"  # TODO: rename to `sideslope`
-        },
-        "waterbody_null_code": -9999,
-        "terminal_code": 0,
-        "waterbody_null_code": -9999,
-        "driver_string": "NetCDF",
-        "layer_string": 0
-    }
-    return params
+    usgs_ind_test = pd.read_csv(io.StringIO(usgs_ind_c),index_col=0).squeeze("columns")
 
-@pytest.fixture
-def waterbody_parameters():
-    return {}
+    gages_c = 'id,value,hydroseq,lake_id\n2624217,08117995,17330.0,\n2624223,TX04138,17356.0,5632093.0\n2624228,08119500,17368.0,\n2624235,08120700,17383.0,\n2624239,08121000,17391.0,\n2624254,08123850,17803.0,\n2624266,TX03517,17856.0,5722027.0\n2624267,08124000,17864.0,\n2624286,08126380,17998.0,\n2624289,08126500,18005.0,\n2624302,TX06386,18791.0,\n2624305,08136700,18797.0,\n2624323,08138000,18903.0,\n2624350,08147000,19841.0,\n2624372,TX00989,19985.0,5756886.0\n2624374,TX00988,19994.0,5756912.0\n2624385,TX00986,21341.0,5756952.0\n2624389,TX00987,21362.0,120053393.0\n2624409,TX01087,21762.0,120053393.0\n2624410,08154510,21763.0,\n2624421,TX01088,21807.0,5781691.0\n2624423,08158000,21810.0,\n2624441,08159200,21944.0,\n2624450,08159500,22090.0,\n2624462,08160400,22161.0,\n2624473,08160700,22215.0,\n2624476,08161000,22319.0,\n2624499,08162000,22383.0,\n2624508,08162500,22435.0,\n2624543,08150000,20555.0,\n2624563,08150700,20928.0,\n2624575,TX03906,21113.0,\n2624577,08151500,21135.0,\n2624610,TX06311,19373.0,\n2624611,08144500,19374.0,\n2624632,08144600,19476.0,\n2624650,08146000,19816.0,\n2624672,TX02243,15999.0,\n2624675,08140700,16031.0,\n2624682,TX02789,16296.0,5734861.0\n2624684,08143500,16324.0,\n2624696,08143600,16437.0,\n2624741,08123650,17646.0,\n2624746,08123720,17655.0,\n2624759,08123800,17722.0,\n2624779,08152900,21598.0,\n2624798,08153500,21660.0,\n2624798,TX01569,21660.0,\n2624824,08128000,18329.0,\n2624831,TX03139,18664.0,5702165.0\n2624832,TX03138,18665.0,\n2624832,TX04415,18665.0,\n2624833,TX03136,18668.0,\n2624834,08136000,18669.0,\n2624835,TX03140,18670.0,\n2624841,TX03142,18676.0,\n2624848,08136500,18753.0,\n2624849,TX01650,18754.0,\n2624871,TX02152,16171.0,5740630.0\n2624874,08140800,16179.0,\n2624894,08144800,19640.0,\n2624907,TX01659,19695.0,5753298.0\n2624909,08145000,19709.0,\n2624941,08148500,20548.0,\n2624960,08129300,18453.0,\n2624962,TX03130,18455.0,\n2624963,08130700,18456.0,\n2624966,08131000,18459.0,5702167.0\n2624968,TX00022,18661.0,5702167.0\n2624970,TX03139,18663.0,5702165.0\n2624978,08133250,18127.0,\n2624984,08133500,18167.0,\n2625000,08134000,18233.0,\n2625003,08134250,18257.0,\n2625007,TX00012,18270.0,120053389.0\n2625008,08135000,18271.0,\n2625010,TX03136,18273.0,\n2625031,08128400,18646.0,\n2625037,08128500,18657.0,5702167.0\n2625061,TX06028,17638.0,5700210.0\n2625095,08152000,21326.0,\n2625107,08158700,15601.0,\n2625111,08158800,15610.0,\n2625117,08159000,15651.0,\n2625227,TX06356,22054.0,\n2625240,08150800,20080.0,\n2625253,08125500,16921.0,5721267.0\n2625253,TX03516,16921.0,5721267.0\n2625302,08130500,18384.0,\n2625304,TX03130,18389.0,\n2625316,08127000,16869.0,\n2625370,08155200,15688.0,\n2625372,08155240,15694.0,\n2625373,08155300,15697.0,\n2625374,08155400,15698.0,\n2625486,TX00006,16229.0,5741244.0\n2625487,08141500,16231.0,\n2625490,08142000,16242.0,\n2625513,08159165,21932.0,\n2625514,08159170,21933.0,\n2625570,08123000,17437.0,5635571.0\n2625571,TX01693,17438.0,5635571.0\n2625649,08120500,17070.0,\n2625684,08158100,21821.0,\n2625684,TX07218,21821.0,\n2625685,08158200,21823.0,\n2625687,08158600,21826.0,\n2625726,TX06305,17563.0,\n2625749,TX01691,17021.0,5635591.0\n2625845,08154700,21784.0,\n2625897,TX02243,15905.0,\n2625902,08139500,18922.0,\n2625965,TX03906,21131.0,\n2625977,TX01659,19585.0,5753298.0\n2625994,TX06311,19143.0,\n2626239,08134230,18255.0,\n2626451,08158819,15625.0,\n2626481,08145000,19712.0,\n2626678,08158860,15637.0,\n2626743,08145000,19708.0,\n2626747,08144600,19487.0,\n2626861,08123755,17666.0,\n2626861,TX03798,17666.0,\n2626900,08159500,22088.0,\n2626928,08158030,15659.0,\n2626929,08158000,15662.0,\n2626947,08158810,15623.0,\n2627067,TX02152,16166.0,5740630.0\n2627087,TX02243,16001.0,\n2627163,08131400,18350.0,\n2627186,08124000,17863.0,\n2627280,08159170,21913.0,\n2627295,08159500,15519.0,\n2627312,08160700,22220.0,\n2627353,TX06902,15644.0,\n2627354,08158970,15648.0,\n2627369,08150800,20011.0,\n2627598,TX06386,16713.0,\n2627628,TX03142,18062.0,\n2627697,TX03517,17845.0,5722027.0\n2627762,08160800,22311.0,\n2627786,08159170,21914.0,\n2627858,08156675,21800.0,\n2627860,08156800,21802.0,\n2627862,TX01088,21804.0,5781691.0\n2627876,08158700,15600.0,\n2627877,08158100,21816.0,\n2627877,TX07218,21816.0,\n2627978,08145000,19575.0,\n2628056,TX00006,16218.0,5741244.0\n2628058,TX00006,16221.0,5741244.0\n2628146,08140500,18925.0,\n2628205,TX01650,18037.0,\n2628214,TX03140,18074.0,\n2628218,TX03142,18061.0,\n2628262,TX03138,18667.0,\n2628262,TX04415,18667.0,\n2628263,TX00022,18462.0,5702167.0\n2628442,08159500,22089.0,\n2628447,TX04380,22177.0,\n2628477,08159200,21945.0,\n2628479,08159165,21916.0,\n2628577,08153500,21661.0,\n2628577,TX01569,21661.0,\n2628583,08158380,21813.0,\n2628584,08158600,21814.0,\n2628586,08154700,21786.0,\n2628630,08155541,21796.0,\n2628633,TX01088,21795.0,5781691.0\n2628634,TX01088,21794.0,5781691.0\n2628647,TX01088,21806.0,5781691.0\n2628819,08148500,20551.0,\n2628941,08144500,19375.0,\n2628998,TX02152,16170.0,5740630.0\n2629009,08143600,16439.0,\n2629108,08133900,18228.0,\n2629155,08128000,18331.0,\n2629188,TX03517,17849.0,5722027.0\n2629190,TX03517,17847.0,5722027.0\n2629326,08159165,21912.0,\n2629540,08158200,21822.0,\n2629541,08158600,21825.0,\n2629543,08158000,15658.0,\n2629544,08158045,15660.0,\n2629545,08158000,15661.0,\n2629551,08154700,21783.0,\n2629596,08155300,15696.0,\n2629597,08155240,15695.0,\n2629599,08155200,15689.0,\n2629620,08158810,15622.0,\n2629625,08158927,15645.0,\n2629626,TX06902,15646.0,\n2629627,08158970,15647.0,\n2629628,08158970,15649.0,\n2629636,08158700,15579.0,\n2629668,08150800,20079.0,\n2629693,08150700,20927.0,\n2629847,08150700,20926.0,\n2630271,08144600,19488.0,\n2630435,08144600,19475.0,\n2630459,08141500,16230.0,\n2630464,08140800,16178.0,\n2630538,08143500,16323.0,\n2630612,TX06386,18790.0,\n2630690,08138000,18902.0,\n2630727,08136500,18752.0,\n2630897,08130500,18385.0,\n2631093,08123650,17647.0,\n'
 
-@pytest.fixture
-def forcing_parameters():
-    return {
-        "nexus_input_folder": _workdir.joinpath("data"),
-        "nexus_file_pattern_filter": "nex-*"
-    }
+    gages_df_test = pd.read_csv(io.StringIO(gages_c),index_col=0)
 
-@pytest.fixture
-def restart_parameters():
-    return {}
-
-@pytest.fixture
-def network(request, waterbody_parameters, restart_parameters, forcing_parameters):
-    type = request.param
-    return HYFeaturesNetwork(supernetwork_parameters(type), waterbody_parameters, restart_parameters, forcing_parameters, verbose=True, showtiming=False)
-
-@pytest.mark.parametrize("network",["gpkg", "json"], indirect=True)
-def test_init(network):
-    #This isn't really nessicary since init is done by the fixture
-    #errors in init will be caught and shown in the test
-    #this test just gives a clear indication that inititalization
-    #is working, instead of infering it from non failure of the fixture creation
-    pass
-
-#@pytest.skip #compute_nhd_routing_v02 has some additonal args that need to be considered
-@pytest.mark.parametrize("network",["gpkg", "json"], indirect=True )
-def test_routable(network, waterbody_parameters):
-        #This is really more of an integration test
-        #but it is here to ensure that the data structure works as intended
-        #when passed off to the router
-        compute_func = "V02-structured"
-
-        results = compute_nhd_routing_v02(
-        network.connections,
-        network.reverse_network,
-        network.waterbody_connections,
-        network.reaches_by_tailwater,
-        compute_func,
-        "by-network",
-        1,
-        # The default here might be the whole network or some percentage...
-        None,
-        network.qlateral.iloc[0].index[0], #t0???
-        300, #dt
-        40, #nts
-        12, #qts_subs
-        network.independent_networks,
-        network.dataframe,
-        network.q0,
-        network.qlateral,
-        pd.DataFrame(),
-        pd.DataFrame(),
-        pd.DataFrame(),
-        pd.DataFrame(),
-        pd.DataFrame(),
-        pd.DataFrame(),
-        {},
-        True,
-        False,
-        network.waterbody_dataframe,
-        waterbody_parameters,  # TODO: Can we remove the dependence on this input? It's like passing argv down into the compute kernel -- seems like we can strip out the specifically needed items.
-        network.waterbody_types_dataframe,
-        not network.waterbody_types_dataframe.index.empty,
-        {},
-    )
+    idx_id_test = gages_df_test.index.name
+    result = {'gages': {2625107: '08158700', 2625111: '08158800', 2626947: '08158810', 2626451: '08158819', 2626678: '08158860', 2629625: '08158927', 2629628: '08158970', 2625117: '08159000', 2626928: '08158030', 2629544: '08158045', 2629599: '08155200', 2629597: '08155240', 2625373: '08155300', 2625374: '08155400', 2624675: '08140700', 2624874: '08140800', 2625487: '08141500', 2625490: '08142000', 2624684: '08143500', 2629009: '08143600', 2625316: '08127000', 2625253: '08125500', 2625649: '08120500', 2624217: '08117995', 2624228: '08119500', 2624235: '08120700', 2624239: '08121000', 2625570: '08123000', 2631093: '08123650', 2624746: '08123720', 2626861: '08123755', 2624759: '08123800', 2624254: '08123850', 2624267: '08124000', 2624286: '08126380', 2624289: '08126500', 2624978: '08133250', 2624984: '08133500', 2629108: '08133900', 2625000: '08134000', 2626239: '08134230', 2625003: '08134250', 2625008: '08135000', 2629155: '08128000', 2627163: '08131400', 2630897: '08130500', 2624960: '08129300', 2624963: '08130700', 2624966: '08131000', 2625031: '08128400', 2625037: '08128500', 2624834: '08136000', 2624848: '08136500', 2624305: '08136700', 2624323: '08138000', 2625902: '08139500', 2628146: '08140500', 2628941: '08144500', 2630271: '08144600', 2624894: '08144800', 2626481: '08145000', 2624650: '08146000', 2624350: '08147000', 2625240: '08150800', 2628819: '08148500', 2624543: '08150000', 2624563: '08150700', 2624577: '08151500', 2625095: '08152000', 2624779: '08152900', 2628577: '08153500', 2624410: '08154510', 2628586: '08154700', 2628630: '08155541', 2627858: '08156675', 2627860: '08156800', 2624423: '08158000', 2628583: '08158380', 2625684: '08158100', 2625685: '08158200', 2625687: '08158600', 2625513: '08159165', 2625514: '08159170', 2628477: '08159200', 2624450: '08159500', 2624462: '08160400', 2627312: '08160700', 2627762: '08160800', 2624476: '08161000', 2624499: '08162000', 2624508: '08162500'}}
+    assert result == hyfn.get_da_gages(gages_df_test, usgs_ind_test, idx_id_test)
